Don't use `hash`.
Between different runs of Python, `hash` is salted differently, so using it
does not preserve ordering between different runs of Python. On the other hand,
IDs are in nearly all cases either ints or strs, both of which can be sorted;
and if the typing system complains, we'll use the _ID types more extensively,
they permit comparison.

Unfortunately, `typing` has no Ordered type or anything similar.