Scriptdoctest does need to run first
Scriptdoctest cannot yet be told to append to an existing coverage report, so it
needs to be run first and other tests need to add to it.

However, when scriptdoctest fails, the other tests should still run, because
they are more informative and verbose.

However however, when only the tour fails, the whole continuous integration test
should still fail.